refactoring project structure: moved payload files
