--- lib/payloads/index.js
@@ -1,11 +1,0 @@
-const credentialsPayload = require('./credentialsPayload');
-const wssLoginPayload = require('./wssLoginPayload');
-const wssUpdatePayload = require('./wssUpdatePayload');
-const zeroConfUpdatePayload = require('./zeroConfUpdatePayload');
-
-module.exports = {
-  credentialsPayload,
-  wssLoginPayload,
-  wssUpdatePayload,
-  zeroConfUpdatePayload,
-};
